+++ platypush/plugins/clipboard/__init__.py
@@ -1,14 +1,27 @@
-from platypush.plugins import Plugin, action
+from typing import Optional
+
+from platypush.context import get_bus
+from platypush.message.event.clipboard import ClipboardEvent
+from platypush.plugins import RunnablePlugin, action
 
 
-class ClipboardPlugin(Plugin):
+class ClipboardPlugin(RunnablePlugin):
     """
-    Plugin to programmatically copy strings to your system clipboard
-    and get the current clipboard content.
+    Plugin to programmatically copy strings to your system clipboard,
+    monitor and get the current clipboard content.
 
     Requires:
-        * **pyperclip** (``pip install pyperclip``)
+        - **pyperclip** (``pip install pyperclip``)
+
+    Triggers:
+
+        - :class:`platypush.message.event.clipboard.ClipboardEvent` on clipboard update.
+
     """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._last_text: Optional[str] = None
 
     @action
     def copy(self, text):
@@ -19,8 +32,8 @@ class ClipboardPlugin(Plugin):
         :type text: str
         """
         import pyperclip
-        pyperclip.copy(text)
 
+        pyperclip.copy(text)
 
     @action
     def paste(self):
@@ -28,8 +41,21 @@ class ClipboardPlugin(Plugin):
         Get the current content of the clipboard
         """
         import pyperclip
+
         return pyperclip.paste()
+
+    def main(self):
+        import pyperclip
+
+        while not self.should_stop():
+            text = pyperclip.paste()
+            if text and text != self._last_text:
+                get_bus().post(ClipboardEvent(text=text))
+                self._last_text = text
+
+            self._should_stop.wait(0.1)
+
+        self.logger.info('Stopped clipboard monitor backend')
 
 
 # vim:sw=4:ts=4:et:
-